refactor: split server.py monolith into 20 domain tool modules
Split the 3295-line server.py into a thin 132-line entry point that
delegates to 20 domain-specific tool modules under src/tools/. All 145
tools preserved with identical names, descriptions, parameters, and
logic. No behavioral changes.
