+++ src/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Tool modules for CallHub MCP Server."""
